+++ src/card/cards/combat_constraints.rs
@@ -210,3 +210,4 @@ pub const DEATHS_PRESENCE: CardDefinitionId = CardDefinitionId(2009);
 pub const AJANI_CALLER_OF_THE_PRIDE: CardDefinitionId = CardDefinitionId(2010);
 pub const SPELL_RUPTURE: CardDefinitionId = CardDefinitionId(2011);
 pub const GIANT_ADEPHAGE: CardDefinitionId = CardDefinitionId(2012);
+pub const KEYMASTER_ROGUE: CardDefinitionId = CardDefinitionId(2013);